Updated value sources to be accurate
The old values I had were overcomplicating and inaccurate for no reason. I thought `client.users` automatically filtered out offline members. It didn't. To remedy that, I instead counted all of the members of each of my guilds, which is inaccurate because one user could be in several mutual guilds.